完成redis同步mysql

+++ ok - 副本/ssm/src/main/java/com/he/ssm/entity/other/Attach.java
@@ -67,5 +67,5 @@ public class Attach {
     @ApiModelProperty(value = "状态")
     private String state;
 
-    private Integer downloadCount;
+    private Long downloadCount;
 }
